Update spirv-tool submodule to include the debug instruct fix (#7541)
* Update spirv-tool submodule to include the debug instruct fix

Update spirv-tool to our forked repo that include the fix of spirv-val
bug. It mistakenly relocates the `DebugFunctionDefinition`.

The upstream PR is: https://github.com/KhronosGroup/SPIRV-Tools/pull/6198
But it's still pending on review.

Issue #7491 will keep track the process, when the upstream PR is merged,
we need to restore the repo back to upstream repo.