Allow a compiler warning on MacOS
This PR is to allow a compiler warning even when all warnings
are requested to be treated as errors.

The following pattern is very common in Slang code base:
  if (auto var = as<...>(...))
And when `var` is not used, the compiler prints a warning.

Alternatively we can remove `auto var =` part but there are too many.