Allow a compiler warning on MacOS (#7561)
This PR is to allow a compiler warning even when all warnings
are requested to be treated as errors.

The following pattern is very common in Slang code base:
  if (auto var = as<...>(...))
And when `var` is not used, the compiler prints a warning.

Alternatively we can remove `auto var =` part but there are too many.

Co-authored-by: Jay Kwak <jkwak@jkwak-mlt.client.nvidia.com>