Debugging code improvement for IR tracking (#7770)
When you track IR creation with `_debugUID` or `_slangIRAllocBreak`,
you had to set a break point near `_slangIRAllocBreak` is used
and unset it after changing the value of `_slangIRAllocBreak`.

Then, when you re-run slangc for more debuggin, you have to set the
break-point again, because it was unset.

With this change, you can keep the break-point in the new line, and
it will hit only once per debugging session.

Not a big improvement, but I found this useful because it requires less
manual handling of the break-points.