slang: Add support for generating getters for IR struct defs. (#7725)
This change expands the IR struct definition generation logic in slang-ir.h.lua
to code generate the getters for the operands of an IR.

To facilitate the above, the schema for the IR definitions in slang-ir-insts.lua is updated
to allow for explicit specification of the operands of an IR, with Fiddle code generating
the getters for them.

slang-ir.h is updated to remove the hardcoded getters of the IRs since they are now generated by Fiddle.

Fixes part of #7185