Run extras/verify-documented-compiler-version.sh only when built
The script, `extras/verify-documented-compiler-version.sh`, depends on
CMakeCache.txt, which gets built as a part of the slang build step.

But if the PR is just a document changes, the build step will be
skipped. And the script will print a warning that CMakeCache.txt is not
found.

The script should run only when Slang was built.