Run extras/verify-documented-compiler-version.sh only when built (#8097)
The script, `extras/verify-documented-compiler-version.sh`, depends on
CMakeCache.txt, which gets built as a part of the slang build step.

But if the PR is just a document changes, the build step will be
skipped. And the script will print a warning that CMakeCache.txt is not
found.

The script should run only when Slang was built.

You can see the current warning here,

https://github.com/shader-slang/slang/actions/runs/16788651517/job/47545476533#step:12:5

Co-authored-by: Ellie Hermaszewska <ellieh@nvidia.com>