fix the cost conversion bug
When we do type coerce, we use a cache to store the conversion cost
of different type. The key of the cache is defined by
struct BasicTypeKey
{
    uint32_t baseType : 8;
    uint32_t dim1 : 4;
    uint32_t dim2 : 4;
    ...
}

where dim1 and dim2 is used for dimension of vector and matrix.
However the dim is only 4 bits, so `vector<int, 16>` will have the same
key as `int`, which is wrong.

Fix the issue by extending it to 8 bit.