Remove the semantic decoration from the original entry struct
When we legalize the entry point param, there are cases where we need to
reconstruct a struct for the parameter and the original struct wouldn't
be used. But if the user tries to use the origianl struct as a type for
a function parameter, we will end up using both the original struct and
the synthesized struct at the same time.

On Metal and WGSL, it causes an error when an identical semtaic is used
on more than one variable.

This commit removes the semantics from the original struct after cloning
the type.