resuce bits for baseType and knownConstantBitCount
Reduce number of bits of baseType to 5. We only have 16 base types,
so it can be represented by only 4 bits, but we need extra 1 bit to
indicate the invalid type.

reduce knownConstantBitCount to 6, because the largest constant number
we can support is 64 bit number, which can be represented by 6 bits.