Remove the comments of embedded source to avoid self-matching
When `SIMPLE` type test is used with `-g[1-3]` option, the filecheck
pattern will most likely to match to the string itself on the embedded
source code rather than match to the emitted spirv-asm code.

This commit avoids the problem by removing the comments in the embedded
source code.

It revealed problems in some of tests:
- tests/bugs/spirv-debug-info.slang : The expected text was missing and
  it had to be added. The file also had Carrage-Return character on all
  lines and the pre-commit git hook removed them.
- tests/spirv/debug-info.slang : the expected keyword `DebugValue` had
  to change to `DebugDeclare`, because that's what we get with ToT.
- tests/spirv/debug-value-dynamic-index.slang : This test is currently
  failing, and it will pass once #7693 is resolved.