Add check for backtrace availability (#8329)
The header execinfo.h and the related backtrace functionality is not
available on all linux platforms. In particular it's missing on musl
linux and on Android before API version 33. This causes compilation
errors on those platforms.

With this change, we first check if backtrace functionality is available
by checking if we are using glibc or a compatible Android version.

Tested on manylinux_2_28 with glibc 2.28 and musllinux_1_2 with musl
1.2, has not been tested on Android.

Co-authored-by: Yong He <yonghe@outlook.com>