Fix SPIRV headers precedence (#8481)
Remove SYSTEM flag from SPIRV-Headers to fix MacOS header precedence.

When the path is registered as SYSTEM, it is used with `-isystem` option
not `-I` option and it gets less searching order on MacOS.

When spirv.h is installed on the system directory, it will end up using
the system installed spirv.h, which is most likely an older version than
we should use.