fix a crash when using type equality constaint
Close #8193.

When constructing `TransitiveTypeWitness` node, we should check
if there is operand that represents two equal times. Currently,
we only check whether the operand is `TypeEqualityWitness`, which
is not good enough, because a `DeclaredSubtypeWitness` could also
be representing two same types, in that case, we should also const
fold this kind of witness.

Fails to do so, we could finally ends up with a generating a
lookup witness IR on a generic parameter that is not supposed to be
looked up.