Remove unnecessary Load and Store pair (#8433)
This commit removes unnecessary Load and Store pairs in IR.

When the IR is like
```
let %1 = var
let %2 = load(%ptr)
store(%1 %2)
```
This PR will replace all uses of %1 with %ptr.
And the load and store instructions will be removed.

But I found that there can be cases where %2 might be still used later
in other IRs.
For these cases, the removal of load instruction relies on DCE.

---------

Co-authored-by: slangbot <ellieh+slangbot@nvidia.com>